requirements.txt
altair==4.2.2
asgiref==3.8.1
attrs==24.3.0
blinker==1.9.0
branca==0.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.8
Django==5.1.3
entrypoints==0.4
exceptiongroup==1.2.2
folium==0.19.4
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.18.4
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pluggy==1.5.0
protobuf==5.29.1
pyarrow==18.1.0
pydeck==0.9.1
Pygments==2.18.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.1
sqlparse==0.5.2
streamlit==1.41.1
streamlit-aggrid==1.0.5
streamlit_folium==0.24.0
tenacity==9.0.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
xyzservices==2024.9.0
psycopg==3.2.3
python-dotenv
djangorestframework
